Refactor smart transaction utility functions
