+++ app/util/smart-transactions/smart-tx.ts
@@ -1,15 +1,11 @@
 import { ApprovalController } from '@metamask/approval-controller';
 import SmartTransactionsController from '@metamask/smart-transactions-controller';
-import {
-  Fee,
-  SmartTransaction,
-} from '@metamask/smart-transactions-controller/dist/types';
+import { SmartTransaction } from '@metamask/smart-transactions-controller/dist/types';
 import {
   TransactionController,
   TransactionMeta,
 } from '@metamask/transaction-controller';
 import Logger from '../Logger';
-import { decimalToHex } from '../conversions';
 import { ApprovalTypes } from '../../core/RPCMethods/RPCMethodMiddleware';
 import TransactionTypes from '../../core/TransactionTypes';
 import {
@@ -18,9 +14,7 @@ import {
   getIsSwapTransaction,
   getIsNativeTokenTransferred,
 } from '../transactions';
-
-// It has to be 21000 for cancel transactions, otherwise the API would reject it.
-const CANCEL_GAS = 21000;
+import { createSignedTransactions } from './utils';
 
 // TODO import these from tx controller
 export declare type Hex = `0x${string}`;
@@ -40,49 +34,6 @@ export interface TransactionParams {
   estimateGasError?: string;
   type?: string;
 }
-
-export const createSignedTransactions = async (
-  unsignedTransaction: TransactionParams,
-  fees: Fee[],
-  areCancelTransactions: boolean,
-  transactionController: TransactionController,
-) => {
-  const unsignedTransactionsWithFees = fees.map((fee) => {
-    const unsignedTransactionWithFees = {
-      ...unsignedTransaction,
-      maxFeePerGas: decimalToHex(fee.maxFeePerGas).toString(),
-      maxPriorityFeePerGas: decimalToHex(fee.maxPriorityFeePerGas).toString(),
-      gas: areCancelTransactions
-        ? decimalToHex(CANCEL_GAS).toString()
-        : unsignedTransaction.gas?.toString(),
-      value: unsignedTransaction.value,
-    };
-    if (areCancelTransactions) {
-      unsignedTransactionWithFees.to = unsignedTransactionWithFees.from;
-      unsignedTransactionWithFees.data = '0x';
-    }
-
-    return unsignedTransactionWithFees;
-  });
-
-  Logger.log(
-    'STX createSignedTransactions, unsignedTransactionsWithFees',
-    unsignedTransactionsWithFees,
-  );
-
-  const signedTransactions =
-    await transactionController.approveTransactionsWithSameNonce(
-      unsignedTransactionsWithFees,
-      { hasNonce: true },
-    );
-
-  Logger.log(
-    'STX createSignedTransactions signedTransactions',
-    signedTransactions,
-  );
-
-  return signedTransactions;
-};
 
 interface Request {
   transactionMeta: TransactionMeta;
